Add: 결제통신준비완료
