+++ src/pages/Payment/components/OrderList.js
@@ -8,11 +8,14 @@ const OrderList = ({ orderInfo }) => {
     <div className="order-list">
       <div className="text-xl">주문제품</div>
       <div className="order-list-main">
-        <div>
-          {orderInfo[0]?.bookTitle}{' '}
-          {orderLength && <span>외 {orderLength}개 </span>}
-          제품을 주문합니다.
-        </div>
+        {orderInfo.length === 0 && <div>장바구니를 불러올 수 없습니다.</div>}
+        {orderInfo.length !== 0 && (
+          <div>
+            {orderInfo[0]?.bookTitle}{' '}
+            {orderLength && <span>외 {orderLength}개 </span>}
+            제품을 주문합니다.
+          </div>
+        )}
       </div>
     </div>
   );
